requirements.txt
Flask==2.3.3
pandas==2.3.1
pdfminer.six==20221105
scikit-learn==1.7.1
spacy==3.7.4
numpy==2.1.3
